test: adds hash compat tests for removing things from shards
License: MIT
